Move Search-related files into Gui.
