+++ src/gui/search/engineselectdlg.cpp
