+++ src/gui/search/engineselectdlg.h
