+++ src/gui/search/pluginsourcedlg.h
