+++ src/gui/search/pluginsourcedlg.ui
