+++ src/gui/search/searchlistdelegate.h
