+++ src/gui/search/searchtab.cpp
