+++ src/gui/search/searchtab.h
