+++ src/gui/search/searchwidget.cpp
