+++ src/gui/search/searchwidget.h
