+++ src/gui/search/searchwidget.ui
