+++ src/gui/search/supportedengines.h
